first commit
Jsw One Banking Based System is a web based application that can be used in any system to allow users to sign up and log in using static password.

In this normal authentication systems, static passwords are used by users to get in the system. after getting in the system the user can perform 4 banking functionalities. and exit the system by clicking on an form input button which has exit written on it.